making npz files

+++ addcoltoDESlightcurve.py
@@ -147,7 +147,7 @@ def addtolightcurve(lightcurvefile,saveloc,mjd,flux,fluxerr,zpt,zptrms,chisq,sky
                     #print line
                     #raw_input()
                     #else:
-                    tidobs = float(line.split()[1])
+                    tidobs = int(line.split()[1])
                     #print tidobs,idobs
                     #raw_input()
                     if int(tidobs) in idobs:
